+++ ZAPD/ZPrerender.h
@@ -23,7 +23,7 @@ public:
 	// static ZPrerender* BuildFromXML(tinyxml2::XMLElement* reader, std::string inFolder,
 	// ZFile* nParent, bool readFile);
 
-	int GetRawDataSize() override;
+	size_t GetRawDataSize() override;
 
 	void DeclareVar(const std::string& prefix, const std::string& bodyStr);
