+++ ZAPD/ZRoom/Commands/ZRoomCommandUnk.h
@@ -18,7 +18,7 @@ public:
 	virtual std::string GenerateSourceCodePass2(std::string roomName, int baseAddress);
 	virtual std::string GenerateSourceCodePass3(std::string roomName);
 	virtual RoomCommand GetRoomCommand();
-	virtual int32_t GetRawDataSize();
+	virtual size_t GetRawDataSize();
 	virtual std::string GetCommandCName();
 	virtual std::string GenerateExterns();
